backend/requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
joblib==1.4.2
nltk==3.9.1
python-dotenv==1.0.1
regex==2024.9.11
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.2
